Remove `tensor_impl::to_row_major_layout`
